+++ TRAP/src/Graphics/API/Vulkan/Objects/VulkanInits.h
@@ -313,7 +313,7 @@ namespace TRAP::Graphics::API::VulkanInits
 	/// <param name="height">Height of the image.</param>
 	/// <param name="depth">Depth of the image.</param>
 	/// <param name="layers">Number of Vulkan image subresource layers.</param>
-	VkBufferImageCopy ImageCopy(uint32_t bufferOffset, uint32_t bufferRowLength, uint32_t bufferImageHeight,
+	VkBufferImageCopy ImageCopy(uint64_t bufferOffset, uint32_t bufferRowLength, uint32_t bufferImageHeight,
 							    uint32_t width, uint32_t height, uint32_t depth, VkImageSubresourceLayers layers) noexcept;
